% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afuns.R
\name{AAfreq}
\alias{AAfreq}
\title{Compute amino acid frequence in TCR repertoire.}
\usage{
AAfreq(Subject.ID, AAseq, Abundance)
}
\arguments{
\item{Subject.ID}{a vector of subject IDs.}

\item{AAseq}{A vector of amino acid sequences. The order of sequences should be consistent with \code{Subject.ID}.}

\item{Abundance}{A numeric vector of abundance. The order of abundances should be consistent with  \code{Subject.ID} and \code{AAseq}.}
}
\value{
\describe{\item{out}{an amino acid matrix with each row represents a subject and each column represents an animo acid.}}
}
\description{
Compute amino acid frequence in TCR repertoire.
}
\examples{
data("Example.data")

## extract features
fR <- AAfreq(TCRdat[,1],TCRdat[,2],as.numeric(TCRdat[,3]))
}
