% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afuns.R
\name{seqhom}
\alias{seqhom}
\title{Compute TCR repertoire homology between subjects.}
\usage{
seqhom(Subject.ID, AAseq, Abundance, substitutionMatrix)
}
\arguments{
\item{Subject.ID}{a vector of subject IDs.}

\item{AAseq}{A vector of amino acid sequences. The order of sequences should be consistent with \code{Subject.ID}.}

\item{Abundance}{A numeric vector of abundance. The order of abundances should be consistent with  \code{Subject.ID} and \code{AAseq}.}

\item{substitutionMatrix}{A character, representing the fixe substitution scores for an alignment. It can take value from following options: BLOSUM45, BLOSUM50, BLOSUM62, BLOSUM80, BLOSUM100, PAM30, PAM40, PAM70, PAM120, and PAM250.}
}
\value{
\describe{\item{S}{a homology matrix.}}
}
\description{
Compute TCR repertoire homology between subjects.
}
\examples{
data("Example.data")

## homology matrix
S <- seqhom(TCRdat[,1],TCRdat[,2],as.numeric(TCRdat[,3]),'BLOSUM62')
}
