% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{tcrl}
\alias{tcrl}
\title{tcrl: Association analysis of T-cell receptor receptor repertoire and clinical phenotypes}
\description{
This package provides a sequence of functions for analyzing the link between the T-cell receptor receptor repertoire and clinical phenotypes.
}
