% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRL.R
\name{TCRL_bin}
\alias{TCRL_bin}
\alias{TCRL_cont}
\title{Score test for TCR repertorire and phenotypes}
\usage{
TCRL_bin(Y, X, fR, W, S)

TCRL_cont(Y, X, fR, W, S)
}
\arguments{
\item{Y}{a vector, \code{Y} should be continous or binary variable.}

\item{X}{a covariate matrix, each row represents a subject.}

\item{fR}{a variant matrix, each row represents a subject, each column represents an extracted feature of TCR repertorire.}

\item{W}{a feature character matrix, each row represents an extracted feature, each column represents a variant. The row number of \code{W} should be consistent with the column number of \code{fR}.}

\item{S}{a homology matrix, which represents the correlation between subjects. It can be cauculated by using function \code{seqhom}.}
}
\value{
\describe{
\item{fix.effect}{p-value for testing the fixed variant effect.}
\item{random.effect}{p-value for testing the random variant effect.}
\item{Overall.pval}{overall p-value for testing the assocation between TCR repertorire and phenotype. It combines \code{fix.effect} and \code{rand.effect} by using Fisher's procedure.}
}
}
\description{
Score test for TCR repertorire and phenotypes
}
\examples{
data("Example.data")
## extract features
fR <- AAfreq(TCRdat[,1],TCRdat[,2],as.numeric(TCRdat[,3]))

## homology matrix
S <- seqhom(TCRdat[,1],TCRdat[,2],as.numeric(TCRdat[,3]),'BLOSUM62')

## include both fixed and random effect
TCRL.cont(Y,X,fR,W,S)
}
