% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRseq.R
\name{TCRseq_bin}
\alias{TCRseq_bin}
\alias{TCRseq_cont}
\title{Score test for TCR repertorire and phenotypes}
\usage{
TCRseq_bin(Y, X, S)

TCRseq_cont(Y, X, S)
}
\arguments{
\item{Y}{a vector, \code{Y} should be continous variable.}

\item{X}{a covariate matrix, each row represents a subject.}

\item{S}{a homology matrix, which represents the correlation between subjects. It can be cauculated by using function \code{seqhom}.}
}
\value{
\item{Overall.pval}{p-value for testing the assocation between TCR repertorire and phenotype.}
}
\description{
Score test for TCR repertorire and phenotypes
}
\examples{
data("Example.data")

## extract features
fR <- AAfreq(TCRdat[,1],TCRdat[,2],as.numeric(TCRdat[,3]))

## homology matrix
S <- seqhom(TCRdat[,1],TCRdat[,2],as.numeric(TCRdat[,3]),'BLOSUM62')

## include random effect only
TCRseq.cont(Y,X,S)
}
